+++ src/gamspy/_symbols/equation.py
@@ -18,6 +18,7 @@ from gams.transfer._internals import (
 import gamspy as gp
 import gamspy._algebra.condition as condition
 import gamspy._algebra.expression as expression
+import gamspy._algebra.operable as operable
 import gamspy._symbols.implicits as implicits
 import gamspy._validation as validation
 import gamspy.utils as utils
@@ -273,7 +274,7 @@ class Equation(gt.Equation, Symbol):
                     " domains are equal"
                 )
 
-            if self.domain_forwarding != domain_forwarding:
+            if self._domain_forwarding != domain_forwarding:
                 raise ValueError(
                     "Cannot overwrite symbol in container unless"
                     " 'domain_forwarding' is left unchanged"
@@ -287,8 +288,8 @@ class Equation(gt.Equation, Symbol):
 
             previous_state = self.container._options.miro_protect
             self.container._options.miro_protect = False
-            self.records = None
-            self.modified = True
+            self._records = None
+            self._modified = True
             self._init_definition(definition)
 
             # only set records if records are provided
@@ -360,14 +361,14 @@ class Equation(gt.Equation, Symbol):
                 self.setRecords(records, uels_on_axes=uels_on_axes)
             else:
                 if not self._is_miro_output:
-                    self.modified = False
+                    self._modified = False
                 self.container._synch_with_gams()
 
             container._options.miro_protect = previous_state
 
     def _serialize(self) -> dict:
         info = {
-            "_domain_forwarding": self.domain_forwarding,
+            "_domain_forwarding": self._domain_forwarding,
             "_is_miro_output": self._is_miro_output,
             "_metadata": self._metadata,
             "_synchronize": self._synchronize,
@@ -553,6 +554,37 @@ class Equation(gt.Equation, Symbol):
 
         self.container._add_statement(statement)
         self._definition = statement
+
+    def _check_ambiguity(self) -> None:
+        """Ambiguity check for MCP, EMP, MPEC models. See #610"""
+        # Looks for =e=, =l= and =g= in an equation definition
+        # with a stack based inorder traversal algorithm (O(N)).
+        stack = []
+
+        assert self._definition is not None
+        node = self._definition.right
+        while True:
+            if node is not None:
+                stack.append(node)
+                node = getattr(node, "left", None)  # type: ignore
+            elif stack:
+                node = stack.pop()
+                if (
+                    isinstance(node, expression.Expression)
+                    and node.operator in {"=e=", "=l=", "=g=", "=x=", "=n="}
+                    and not isinstance(node.right, operable.Operable)
+                ):
+                    raise ValidationError(
+                        f"Definition of `{self.name}` is ambigiuous. Please "
+                        "use gp.Number for numeric values or disable ambiguity "
+                        "check via gp.set_options({'ALLOW_AMBIGUOUS_EQUATIONS': 'no'}). "
+                        "Using numeric values in equations without gp.Number can result in "
+                        f"different order than expected. Print `{self.name}.getDefinition()` to "
+                        "make sure that the equation definition is as expected."
+                    )
+                node = getattr(node, "right", None)
+            else:
+                break  # pragma: no cover
 
     @property
     def l(self):  # noqa: E741, E743
@@ -1000,12 +1032,12 @@ class Equation(gt.Equation, Symbol):
         self._records = records
 
         self._requires_state_check = True
-        self.modified = True
+        self._modified = True
 
         self.container._requires_state_check = True
         self.container.modified = True
 
-        if self._records is not None and self.domain_forwarding:
+        if self._records is not None and self._domain_forwarding:
             self._domainForwarding()
 
             # reset state check flags for all symbols in the container
@@ -1159,7 +1191,7 @@ class Equation(gt.Equation, Symbol):
         output = f"Equation {self.name}"
 
         if self.domain:
-            output += self._get_domain_str(self.domain_forwarding)
+            output += self._get_domain_str(self._domain_forwarding)
 
         if self.description:
             output += ' "' + self.description + '"'