+++ src/gamspy/formulations/ml/__init__.py
@@ -1,4 +1,11 @@
 from gamspy.formulations.ml.decision_tree_struct import DecisionTreeStruct
+from gamspy.formulations.ml.gradient_boosting import GradientBoosting
+from gamspy.formulations.ml.random_forest import RandomForest
 from gamspy.formulations.ml.regression_tree import RegressionTree
 
-__all__ = ["RegressionTree", "DecisionTreeStruct"]
+__all__ = [
+    "RegressionTree",
+    "DecisionTreeStruct",
+    "RandomForest",
+    "GradientBoosting",
+]