+++ src/gamspy/formulations/nn/conv1d.py
@@ -1,12 +1,13 @@
 from __future__ import annotations
 
+import math
 from typing import TYPE_CHECKING, Literal
 
 import numpy as np
 from numpy.lib.stride_tricks import sliding_window_view
 
 import gamspy as gp
-import gamspy.formulations.nn.utils as utils
+import gamspy.formulations.utils as utils
 from gamspy.exceptions import ValidationError
 from gamspy.math import dim
 
@@ -300,14 +301,21 @@ class Conv1d:
         output.lo[...] = out_bounds[("0",) + tuple(output.domain)]
         output.up[...] = out_bounds[("1",) + tuple(output.domain)]
 
-    def make_variable(self) -> None:
+    def make_variable(self, *, init_weights=False) -> None:
         """
         Mark Conv1d as variable. After this is called `load_weights`
         cannot be called. Use this when you need to learn the weights
         of your convolutional layer in your optimization model.
 
-        This does not initialize the weights, it is highly recommended
-        that you set initial values to `weight` and `bias` variables.
+        Parameters
+        ----------
+        init_weights : Optional[bool]
+               False by default.
+               Whether to initialize weights. It is suggested you set
+               this to True unless you want to initialize weights yourself.
+               When `init_weights` is set to True, values are initialized from
+               :math:`\\mathcal{U}(-\\sqrt{k},\\sqrt{k})`, where :math:`k = 1/[C_{in} * kernel\\_size]`.
+
         """
         if self._state == 1:
             raise ValidationError(
@@ -320,12 +328,15 @@ class Conv1d:
             self.kernel_size,
         )
 
+        sk = math.sqrt(1 / (self.in_channels * self.kernel_size))
         if self.weight is None:
             self.weight = gp.Variable(
                 self.container,
                 domain=dim(expected_shape),
                 name=utils._generate_name("v", self._name_prefix, "weight"),
             )
+            if init_weights:
+                self.weight.l[...] = gp.math.uniform(-sk, sk)
 
         if self.use_bias and self.bias is None:
             self.bias = gp.Variable(
@@ -333,6 +344,8 @@ class Conv1d:
                 domain=dim([self.out_channels]),
                 name=utils._generate_name("v", self._name_prefix, "bias"),
             )
+            if init_weights:
+                self.bias.l[...] = gp.math.uniform(-sk, sk)
 
         self._state = 2