+++ tests/unit/test_nn_formulation.py
@@ -247,6 +247,22 @@ def test_conv_make_variable(data):
             padding=padding_type,
         )
         conv1.make_variable()
+        assert conv1.weight.records is None
+        assert conv1.bias.records is None
+
+        conv2 = layer(
+            m,
+            in_channels=1,
+            out_channels=2,
+            kernel_size=3,
+            bias=True,
+            padding=padding_type,
+        )
+        # setting init_weights initializes bias and weight
+        conv2.make_variable(init_weights=True)
+        assert conv2.weight.records is not None
+        assert conv2.bias.records is not None
+
         w1 = np.random.rand(*[2, 1, 3, 3][:shape_len])
         b1 = np.random.rand(2)
         pytest.raises(ValidationError, conv1.load_weights, w1, b1)
@@ -2943,6 +2959,13 @@ def test_linear_make_variable(data):
     m, *_ = data
     lin1 = Linear(m, 4, 2)
     lin1.make_variable()
+    assert lin1.weight.records is None
+    assert lin1.bias.records is None
+
+    lin2 = Linear(m, 4, 2)
+    lin2.make_variable(init_weights=True)
+    assert lin2.weight.records is not None
+    assert lin2.bias.records is not None
     w1 = np.random.rand(2, 4)
     b1 = np.random.rand(2)
     pytest.raises(ValidationError, lin1.load_weights, w1, b1)
@@ -3178,9 +3201,8 @@ def test_linear_propagate_unbounded_input_with_zero_weight(data):
     x = gp.Variable(m, "x", domain=dim([30, 20, 30, 20]))
 
     out1, _ = lin1(x)
-
-    out1_ub = np.array(out1.up.records.upper).reshape(30, 20, 30, 30)
-    out1_lb = np.array(out1.lo.records.lower).reshape(30, 20, 30, 30)
+    out1_ub = np.array(out1.records.upper).reshape(30, 20, 30, 30)
+    out1_lb = np.array(out1.records.lower).reshape(30, 20, 30, 30)
 
     expected_bounds = np.zeros((30, 20, 30, 30))